Fix memory leaks in macros and improve makefile
Add calls to free in debug prints that do not run because debugging is
currently disabled. Remove macros for printing arrays because they
leaked memory. The makefile now checks for an environment variable
called "DEBUG" that is set to "true" and if this is the case then make
will append "-D DEBUG" to the flags passed to the compiler.